android: conversion to new Android Publisher, initial batch
If these jobs make it, then the rest will do too.

+++ android-lcr/hikey/builders.sh
@@ -1,5 +1,11 @@
-# Build Android
-## clean android-patchsets and repositories in device
+# Early test
+if [ ! -f build-configs/${BUILD_CONFIG_FILENAME} ]; then
+  echo "No config file named ${BUILD_CONFIG_FILENAME} exists"
+  echo "in android-build-configs.git"
+  exit 1
+fi
+
+# Clean android-patchsets and repositories in device
 rm -rf build/out build/android-patchsets build/device
 
 mkdir -p build/
@@ -9,6 +15,7 @@ tar -xvf linaro-hikey-20160226-67c37b1a.tgz
 yes "I ACCEPT" | ./extract-linaro-hikey.sh
 cd -
 
+# Build Android
 build-tools/node/build us-east-1.ec2-git-mirror.linaro.org "${CONFIG}"
 cp -a /home/buildslave/srv/${BUILD_DIR}/build/out/*.json /home/buildslave/srv/${BUILD_DIR}/build/out/*.xml ${WORKSPACE}/
 
@@ -28,29 +35,14 @@ cd -
 rm -rf build/out/BUILD-INFO.txt
 wget https://git.linaro.org/ci/job/configs.git/blob_plain/HEAD:/android-lcr/hikey/build-info/template.txt -O build/out/BUILD-INFO.txt
 
-# Publish binaries
-PUB_DEST=/android/$JOB_NAME/$BUILD_NUMBER
-time linaro-cp.py \
-  --api_version 3 \
-  --manifest \
-  --no-build-info \
-  --link-latest \
-  --split-job-owner \
-  build/out \
-  ${PUB_DEST} \
-  --include "^[^/]+[._](img[^/]*|tar[^/]*|xml|sh|config)$" \
-  --include "^[BHi][^/]+txt$" \
-  --include "^(MANIFEST|MD5SUMS|changelog.txt)$"
+# Publish parameters
+cat << EOF > ${WORKSPACE}/publish_parameters
+PUB_SRC=${PWD}/build/out
+PUB_DEST=/android/${JOB_NAME}/${BUILD_NUMBER}
+EOF
 
 # Construct post-build-lava parameters
-if [ -f build-configs/${BUILD_CONFIG_FILENAME} ]; then
-  source build-configs/${BUILD_CONFIG_FILENAME}
-else
-  echo "No config file named ${BUILD_CONFIG_FILENAME} exists"
-  echo "in android-build-configs.git"
-  exit 1
-fi
-
+source build-configs/${BUILD_CONFIG_FILENAME}
 cat << EOF > ${WORKSPACE}/post_build_lava_parameters
 DEVICE_TYPE=${LAVA_DEVICE_TYPE:-${TARGET_PRODUCT}}
 TARGET_PRODUCT=${TARGET_PRODUCT}
@@ -65,5 +57,3 @@ DOWNLOAD_URL=${PUBLISH_SERVER}/${PUB_DEST}
 CUSTOM_JSON_URL=https://git.linaro.org/qa/test-plans.git/blob_plain/HEAD:/android/hikey/template.json
 SKIP_REPORT=false
 EOF
-
-echo "Build finished"
